Add API documentation
This commit adds API documentation using apifairy.
