+++ app/__init__.py
@@ -3,6 +3,7 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 from flask_login import LoginManager
 from flask_mail import Mail
+from apifairy import APIFairy
 from elasticsearch import Elasticsearch
 from redis import Redis
 import rq
@@ -11,6 +12,7 @@ from config import config
 db = SQLAlchemy()
 mg = Migrate()
 login = LoginManager()
+fairy = APIFairy()
 login.login_view = 'auth.login'
 login.login_message = 'Please log in to access this page.'
 mail = Mail()
@@ -23,19 +25,20 @@ def create_app(config_name='default'):
 
     # Disable trailing slash 
     app.url_map.strict_slashes = False
-    
+
     # Initialize extensions
     db.init_app(app)
     mg.init_app(app, db)
     login.init_app(app)
     mail.init_app(app)
+    fairy.init_app(app)
     app.elasticsearch = Elasticsearch([app.config['ELASTICSEARCH_URL']]) \
         if app.config['ELASTICSEARCH_URL'] else None
     app.redis = Redis.from_url(app.config['REDIS_URL'])
     app.task_queue = rq.Queue('app-tasks', connection=app.redis)
 
     # Register blueprints
-    from app.cli import bp as cli_bp 
+    from app.cli import bp as cli_bp
     app.register_blueprint(cli_bp)
     from app.main import bp as main_bp
     app.register_blueprint(main_bp)
@@ -43,6 +46,8 @@ def create_app(config_name='default'):
     app.register_blueprint(admin_bp)
     from app.auth import bp as auth_bp
     app.register_blueprint(auth_bp)
+    from app.api import bp as api_bp
+    app.register_blueprint(api_bp, url_prefix='/api')
 
     from app import models
     @app.shell_context_processor
