+++ app/api/documentation.py
@@ -1,0 +1,8 @@
+from flask import redirect, url_for
+
+from app.api import bp
+
+
+@bp.route('/docs')
+def docs():
+    return redirect(url_for('apifairy.docs'))
