+++ app/main/routes.py
@@ -30,26 +30,6 @@ def user():
 def wishlist():
     return current_user.wishlist.products
 
-@bp.route('/cart', methods=['GET'])
-@login_required
-def cart():
-    return current_user.cart.products
-
-
-@bp.route('/cart/<int:product_id>/<int:quantity>', methods=['GET'])
-@login_required
-def cart_quantity(product_id, quantity):
-    # Update the quantity in the database
-    product = Product.query.get(product_id)
-    current_user.cart.update_quantity(product, quantity)
-    db.session.commit()
-    return jsonify({
-        'product_quantity': quantity,
-        'total_price': current_user.cart.total_price,
-        'product_total_price': current_user.cart.cart_product_total_price(product),
-        'product_total_sale_price': current_user.cart.cart_product_total_sale_price(product),
-    })
-
 
 @bp.route('/checkout', methods=['GET'])
 @login_required
