+++ app/templates/navigator.html
@@ -110,6 +110,17 @@
                   <span class="items__count">{{ current_user.cart.cart_products|length }}</span> 
                 </a>
               </li>
+            {% else %}
+              <!-- login button -->
+              <li class="header__account--items">
+                <a class="header__account--btn" href="{{ url_for('auth.login') }}">
+                  <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-box-arrow-in-right" viewBox="0 0 16 16">
+                    <path fill-rule="evenodd" d="M6 3.5a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v9a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-2a.5.5 0 0 0-1 0v2A1.5 1.5 0 0 0 6.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-9A1.5 1.5 0 0 0 14.5 2h-8A1.5 1.5 0 0 0 5 3.5v2a.5.5 0 0 0 1 0v-2z"/>
+                    <path fill-rule="evenodd" d="M11.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H1.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3z"/>
+                  </svg>
+                  <span class="visually-hidden">Login</span> 
+                </a>
+              </li>
             {% endif %}
           </ul>
         </div>
@@ -453,6 +464,19 @@
             <span class="items__count">{{ current_user.wishlist.products|length }}</span> 
           </a>
         </li>
+      {% else %}
+        <!-- login bottom scren small -->
+        <li class="offcanvas__stikcy--toolbar__list">
+          <a class="offcanvas__stikcy--toolbar__btn" href="{{ url_for('auth.login') }}">
+            <span class="offcanvas__stikcy--toolbar__icon"> 
+              <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-box-arrow-in-right" viewBox="0 0 16 16">
+                <path fill-rule="evenodd" d="M6 3.5a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v9a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-2a.5.5 0 0 0-1 0v2A1.5 1.5 0 0 0 6.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-9A1.5 1.5 0 0 0 14.5 2h-8A1.5 1.5 0 0 0 5 3.5v2a.5.5 0 0 0 1 0v-2z"/>
+                <path fill-rule="evenodd" d="M11.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H1.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3z"/>
+              </svg>
+            </span>
+            <span class="offcanvas__stikcy--toolbar__label">Login</span>
+          </a>
+        </li>
       {% endif %}
     </ul>
   </div>
@@ -510,7 +534,7 @@
         </div>
       </div>
       <div class="minicart__button d-flex justify-content-center">
-        <a class="btn minicart__button--link" href="{{ url_for('main.cart') }}">View cart</a>
+        <a class="btn minicart__button--link" href="{{ url_for('api.cart') }}">View cart</a>
         <a class="btn minicart__button--link" href="{{ url_for('main.checkout') }}">Checkout</a>
       </div>
     </div>
