+++ requirements.txt
@@ -1,4 +1,6 @@
 alembic==1.9.3
+apifairy==1.3.0
+apispec==6.3.0
 async-timeout==4.0.2
 blinker==1.5
 certifi==2022.12.7
@@ -10,8 +12,10 @@ elasticsearch==8.6.1
 email-validator==1.3.1
 Faker==16.7.0
 Flask==2.2.2
+Flask-HTTPAuth==4.7.0
 Flask-Login==0.6.2
 Flask-Mail==0.9.1
+flask-marshmallow==0.14.0
 Flask-Migrate==4.0.4
 Flask-SQLAlchemy==3.0.3
 Flask-WTF==1.1.1
@@ -32,5 +36,6 @@ SQLAlchemy==2.0.2
 typing_extensions==4.4.0
 urllib3==1.26.14
 visitor==0.1.3
+webargs==8.2.0
 Werkzeug==2.2.2
 WTForms==3.0.1
